+++ packages/primitive/components/slot/src/utils/merge-props.ts
